Refactor photo gallery and fetch series photos
